Move discovery endpoint package into unaggregated
