+++ staging/src/k8s.io/apiserver/pkg/endpoints/discovery/unaggregated/addresses.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package discovery
+package unaggregated
 
 import (
 	"net"
